diffextract.goをextract.goにリネーム
古いextract.goは使用されてなかったため削除
